+++ libloco/include/Arrays/LocoUShortArray.h
@@ -16,7 +16,7 @@ class UShortArray : public Object {
 	Q_PROPERTY( quint64 length READ GetLength )
     Q_PROPERTY( QString valueType READ GetValueType )
     Q_PROPERTY( QVariantList data READ GetData WRITE SetData )
-    Q_PROPERTY( int alignment READ GetAlignment )
+    Q_PROPERTY( quint64 alignment READ GetAlignment )
     Q_PROPERTY( quint64 address READ Address )
 private:
     unsigned short ToUShort( const QVariant& v ) { return ( unsigned short )( v.toInt() ); }
